Refactor: Writing 페이지 디렉토리 변경
