--- src/components/organisms/writing/index.js
@@ -1,1 +1,0 @@
-export { default as WritingContent } from './WritingContent';
